+++ includes/class-rest-capabilities.php
@@ -495,6 +495,7 @@ class Capabilities extends Base {
 			'rondo_iva_approve'  => 'IVA goedkeuren',
 		];
 		$capability_labels[ \Rondo\Core\UserRoles::KADERLIJST_CAPABILITY ] = 'Kaderlijst';
+		$capability_labels[ \Rondo\Core\UserRoles::TRAINING_CAPABILITY ]   = 'Trainingsschema beheren';
 
 		$wp_roles     = wp_roles();
 		$all_roles    = \Rondo\Core\UserRoles::get_all_roles();
@@ -563,6 +564,7 @@ class Capabilities extends Base {
 
 		$allowed_caps   = [ 'fairplay', 'vog', 'financieel', 'financieel_read', 'toegangscontrole', 'manage_clothing', 'ledenadministratie', 'sponsorbeheer', 'narrowcasting', 'accommodatiebeheer', 'vrijwilligers', 'rondo_iva_approve' ];
 		$allowed_caps[] = \Rondo\Core\UserRoles::KADERLIJST_CAPABILITY;
+		$allowed_caps[] = \Rondo\Core\UserRoles::TRAINING_CAPABILITY;
 		$valid_slugs    = array_keys( \Rondo\Core\UserRoles::get_all_roles() );
 		$valid_slugs[]  = 'administrator';