+++ includes/class-rest-user-settings.php
@@ -1218,6 +1218,7 @@ class UserSettings extends Base {
 			'can_access_narrowcasting'      => \Rondo\Config\FeatureToggles::can_access( 'narrowcasting' ) && ( current_user_can( 'narrowcasting' ) || current_user_can( 'sponsorbeheer' ) || $is_admin ),
 			'can_manage_narrowcasting'      => \Rondo\Config\FeatureToggles::can_access( 'narrowcasting' ) && ( current_user_can( 'narrowcasting' ) || $is_admin ),
 			'can_manage_accommodatie'       => \Rondo\Config\FeatureToggles::can_access( 'rooms' ) && ( current_user_can( 'accommodatiebeheer' ) || $is_admin ),
+			'can_manage_training'           => \Rondo\Core\UserRoles::can_manage_training( $user_id ),
 			'can_manage_tournaments'        => \Rondo\Tournaments\TournamentAccess::can_manage( $user_id ),
 			'has_tournament_assignments'    => \Rondo\Tournaments\TournamentAccess::has_assignments( $user_id ),
 			'has_my_teams'                  => ! empty( \Rondo\Teams\MyTeam::teams_for_user( $user_id ) ),